Ceil instead of floor VotesNeeded for RTV
Since RTV:ing is a destructive action in surf and KZ servers, where you
can destroy a persons run if you RTV, we should ceil instead of floor.
Doing this means that instead of allowing 1 person to RTV when 2 are on
the server, we will require both persons to RTV for a map change, so you
cannot join and ruin someone's run.