Remove NativeVotes
NativeVotes does not seem to work anymore after various updates. It does
not seem to be needed in CSGO, which is the game I am mainly building
this for, so lets get rid of it and see what happens.

Additionally it makes the whole logic with menus a lot simpler, since
there is no longer duplicates around everything.