Replace emitsoundany with using sdktools instead
Apparently after the 2018 csgo update, the emitsoundany include is no
longer needed. So instaed of adding that, we are just using what
Sourcemod provides.

Also setting precache to true on the sounds, so that there are no
unpleasant lags or anything when they are supposed to be used.

Completely untested changes, but it compiles!